Abstract away GSettings
This is a preparation step for #183 i.e. migrating from GSettings 0.1
to 0.2.

GSettings 0.1 and 0.2 are vastly different because the FFI-based 0.1 is
synchronous, whereas the D-Bus-based 0.2 is asynchronous. Furthermore,
0.1 has separate getters for different primitive types, whereas 0.2 has
only one getter that returns a DBusValue.

In order to make the migration from GSettings 0.1 to 0.2 easier, hide
GSettings instances from the models and use nullable setting values to
eliminate the assumption that setting values are immediately available.